Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int a;
-    int b;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    printf("Product: %d",a*b);
+    float a;
+    float b;
+    scanf("%f",&a);
+    scanf("%f",&b);
+    printf("Product: %.2f",a*b);
     return 0;
 }
